backend/requirements.txt
-f https://download.pytorch.org/whl/cpu
fastapi
uvicorn
wandb
python-multipart
numpy==1.26.4
Pillow==11.1.0
python-dotenv==1.1.0
matplotlib==3.9.4
# torch==2.6.0
# torchvision==0.21.0